+++ jest.setupAfterEnv.js
@@ -1,5 +1,7 @@
 "use strict";
 
+process.env.CDS_CONFIG = '{ "requires": { "outbox": "persistent-outbox" } }';
+
 // turn off regular and error logging;
 // jest.spyOn(console, "log").mockImplementation();
 // jest.spyOn(console, "info").mockImplementation();