+++ src/EventQueueProcessorBase.js
@@ -12,7 +12,7 @@ const eventConfig = require("./config");
 const PerformanceTracer = require("./shared/PerformanceTracer");
 
 const IMPLEMENT_ERROR_MESSAGE = "needs to be reimplemented";
-const COMPONENT_NAME = "eventQueue/EventQueueProcessorBase";
+const COMPONENT_NAME = "/eventQueue/EventQueueProcessorBase";
 
 const DEFAULT_RETRY_ATTEMPTS = 3;
 const DEFAULT_PARALLEL_EVENT_PROCESSING = 1;
@@ -131,7 +131,7 @@ class EventQueueProcessorBase {
 
   endPerformanceTracerEvents() {
     this.__performanceLoggerEvents?.endPerformanceTrace(
-      { threshold: 50 },
+      { threshold: this.#config.thresholdLoggingEventProcessing },
       {
         eventType: this.#eventType,
         eventSubType: this.#eventSubType,
@@ -141,7 +141,7 @@ class EventQueueProcessorBase {
 
   endPerformanceTracerPeriodicEvents() {
     this.__performanceLoggerPeriodicEvents?.endPerformanceTrace(
-      { threshold: 50 },
+      { threshold: this.#config.thresholdLoggingEventProcessing },
       {
         eventType: this.#eventType,
         eventSubType: this.#eventSubType,
@@ -151,7 +151,7 @@ class EventQueueProcessorBase {
 
   endPerformanceTracerPreprocessing() {
     this.__performanceLoggerPreprocessing?.endPerformanceTrace(
-      { threshold: 50 },
+      { threshold: this.#config.thresholdLoggingEventProcessing },
       {
         eventType: this.#eventType,
         eventSubType: this.#eventSubType,
@@ -869,7 +869,7 @@ class EventQueueProcessorBase {
     return await checkAndUpdatePromise;
   }
 
-  async handleDistributedLock() {
+  async acquireDistributedLock() {
     if (this.concurrentEventProcessing) {
       return true;
     }