+++ src/initialize.js
@@ -2,7 +2,6 @@
 
 const { promisify } = require("util");
 const fs = require("fs");
-const path = require("path");
 
 const cds = require("@sap/cds");
 const yaml = require("yaml");
@@ -14,6 +13,7 @@ const dbHandler = require("./dbHandler");
 const config = require("./config");
 const { initEventQueueRedisSubscribe, closeSubscribeClient } = require("./redisPubSub");
 const { closeMainClient } = require("./shared/redis");
+const eventQueueAsOutbox = require("./outbox/eventQueueAsOutbox");
 
 const readFileAsync = promisify(fs.readFile);
 
@@ -34,6 +34,9 @@ const CONFIG_VARS = [
   ["disableRedis", false],
   ["skipCsnCheck", false],
   ["updatePeriodicEvents", true],
+  ["thresholdLoggingEventProcessing", 50],
+  ["useAsCAPOutbox", false],
+  ["userId", null],
 ];
 
 const initialize = async ({
@@ -47,6 +50,9 @@ const initialize = async ({
   disableRedis,
   skipCsnCheck,
   updatePeriodicEvents,
+  thresholdLoggingEventProcessing,
+  useAsCAPOutbox,
+  userId,
 } = {}) => {
   // TODO: initialize check:
   // - content of yaml check
@@ -67,20 +73,26 @@ const initialize = async ({
     tableNameEventLock,
     disableRedis,
     skipCsnCheck,
-    updatePeriodicEvents
+    updatePeriodicEvents,
+    thresholdLoggingEventProcessing,
+    useAsCAPOutbox,
+    userId
   );
 
   const logger = cds.log(COMPONENT);
   config.fileContent = await readConfigFromFile(config.configFilePath);
   config.checkRedisEnabled();
 
-  const dbService = await cds.connect.to("db");
-  await (cds.model ? Promise.resolve() : new Promise((resolve) => cds.on("serving", resolve)));
-  !config.skipCsnCheck && (await csnCheck());
   if (config.processEventsAfterPublish) {
-    dbHandler.registerEventQueueDbHandler(dbService);
+    cds.on("connect", (service) => {
+      if (service.name === "db") {
+        dbHandler.registerEventQueueDbHandler(service);
+      }
+    });
   }
+  !config.skipCsnCheck && (await csnCheck());
 
+  monkeyPatchCAPOutbox();
   registerEventProcessors();
   registerCdsShutdown();
   logger.info("event queue initialized", {
@@ -130,27 +142,42 @@ const registerEventProcessors = () => {
   }
 };
 
+const monkeyPatchCAPOutbox = () => {
+  if (config.useAsCAPOutbox) {
+    Object.defineProperty(cds, "outboxed", {
+      get: () => eventQueueAsOutbox.outboxed,
+    });
+    Object.defineProperty(cds, "unboxed", {
+      get: () => eventQueueAsOutbox.unboxed,
+    });
+  }
+};
+
 const csnCheck = async () => {
-  const eventCsn = cds.model.definitions[config.tableNameEventQueue];
-  if (!eventCsn) {
-    throw EventQueueError.missingTableInCsn(config.tableNameEventQueue);
-  }
+  cds.on("loaded", async (csn) => {
+    if (csn.namespace === "cds.xt") {
+      return;
+    }
+    const eventCsn = csn.definitions[config.tableNameEventQueue];
+    if (!eventCsn) {
+      throw EventQueueError.missingTableInCsn(config.tableNameEventQueue);
+    }
 
-  const lockCsn = cds.model.definitions[config.tableNameEventLock];
-  if (!lockCsn) {
-    throw EventQueueError.missingTableInCsn(config.tableNameEventLock);
-  }
+    const lockCsn = csn.definitions[config.tableNameEventLock];
+    if (!lockCsn) {
+      throw EventQueueError.missingTableInCsn(config.tableNameEventLock);
+    }
 
-  if (config.tableNameEventQueue === BASE_TABLES.EVENT && config.tableNameEventLock === BASE_TABLES.LOCK) {
-    return; // no need to check base tables
-  }
+    if (config.tableNameEventQueue === BASE_TABLES.EVENT && config.tableNameEventLock === BASE_TABLES.LOCK) {
+      return; // no need to check base tables
+    }
 
-  const csn = await cds.load(path.join(__dirname, "..", "db"));
-  const baseEvent = csn.definitions["sap.eventqueue.Event"];
-  const baseLock = csn.definitions["sap.eventqueue.Lock"];
+    const baseEvent = csn.definitions["sap.eventqueue.Event"];
+    const baseLock = csn.definitions["sap.eventqueue.Lock"];
 
-  checkCustomTable(baseEvent, eventCsn);
-  checkCustomTable(baseLock, lockCsn);
+    checkCustomTable(baseEvent, eventCsn);
+    checkCustomTable(baseLock, lockCsn);
+  });
 };
 
 const checkCustomTable = (baseCsn, customCsn) => {